% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyAnnotation.R
\name{corpAnnotate}
\alias{corpAnnotate}
\alias{buildTuple}
\alias{ModifyAnnotation}
\title{Modify text annotations for more robust corpus search}
\usage{
buildTuple(x)

ModifyAnnotation(x)
}
\arguments{
\item{x}{A list of dataframes}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
