% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_context_bow.R
\name{clr_context_bow}
\alias{clr_context_bow}
\title{Summarize results of corpuslingr::clr_search_contexts()}
\usage{
clr_context_bow(x, content_only = TRUE, agg_var = c("lemma", "pos"))
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
A function for accessing/aggregating BOW object returned from corpuslingr::clr_search_contexts().
}
