% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_context_kwic.R
\name{clr_context_kwic}
\alias{clr_context_kwic}
\title{Summarize results of corpuslingr::clr_search_contexts()}
\usage{
clr_context_kwic(x, include = c("doc_id", "lemma"))
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
A function for accessing/aggregating BOW object returned from corpuslingr::clr_search_contexts().
}
