% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_desc_corpus.R
\name{clr_desc_corpus}
\alias{clr_desc_corpus}
\title{Provides basic frequency stats for an annotated corpus.}
\usage{
clr_desc_corpus(corp, doc = "id", sent = "sid", tok = "word",
  upos = "upos", genre = NULL)
}
\value{
A dataframe
}
\description{
Summarizes token frequency, type frequency, # of sentences by corpus/text/(optionally) genre.
}
