% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_get_freq.R
\name{clr_get_freq}
\alias{clr_get_freq}
\title{Summarize results of corpuslingr::GetContexts()}
\usage{
clr_get_freq(x, agg_var = c("lemma", "token"), toupper = FALSE)
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
These functions aggregate search results by frequency, BOW, and KWIC.
}
