% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_prep_corpus.R
\name{clr_prep_corpus}
\alias{clr_prep_corpus}
\title{Clean 'text' column of corpus dataframe}
\usage{
clr_prep_corpus(x, hyphenate = TRUE)
}
\arguments{
\item{x}{A dataframe corpus}
}
\value{
A dataframe corpus
}
\description{
Function modifies & cleans 'text' column of corpus dataframe prior to annotation; addresses issues such as long-dashes, hyphenated words, and unwanted white space.
}
