% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_set_corpus.R
\name{clr_set_corpus}
\alias{clr_set_corpus}
\alias{clr_set_tuple}
\title{Modify text annotations for more robust corpus search}
\usage{
clr_set_tuple(x)
}
\arguments{
\item{x}{A list of dataframes}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
