% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpAnnotate.R
\name{corpAnnotate}
\alias{corpAnnotate}
\alias{PrepText}
\alias{buildTuple}
\alias{ModifyAnnotation}
\title{Annotate a corpus of texts using the `spacyr` package}
\usage{
PrepText(x, hyphenate = TRUE)

buildTuple(x)

ModifyAnnotation(x)
}
\arguments{
\item{x}{A list of dataframes}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
