% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpOutput.R
\name{corpOutput}
\alias{corpOutput}
\alias{GetSearchFreqs}
\alias{GetKWIC}
\alias{GetBOW}
\alias{GetDocDesc}
\title{Annotate a corpus of texts using the `spacyr` package}
\usage{
GetSearchFreqs(x, aggBy = "lemma")

GetKWIC(x)

GetBOW(x, contentOnly)

GetDocDesc(x)
}
\arguments{
\item{charList}{A list of texts as character strings}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
