% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpSearch.R
\docType{data}
\name{corpSearch}
\alias{corpSearch}
\alias{buildSearch}
\alias{nounPhrase}
\alias{keyPhrase}
\alias{GetContexts}
\title{Get search in context}
\format{An object of class \code{character} of length 1.}
\usage{
buildSearch(x)

nounPhrase

keyPhrase

GetContexts(search, corp, LW, RW)
}
\arguments{
\item{search}{Gram/lexical pattern to be searched for}

\item{corp}{List of annotated texts to be searched}

\item{LW}{Size of context in number of words to left of the target}

\item{RW}{Size of context in number of words to right of the target}
}
\value{
A list of dataframes
}
\description{
These functions enable corpus search of gram constructions in context.
}
\keyword{datasets}
