% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_corpus.R
\name{describe_corpus}
\alias{describe_corpus}
\alias{clr_desc_corpus}
\title{Provides basic frequency stats for an annotated corpus.}
\usage{
clr_desc_corpus(corp, doc = "id", sent = "sid", tok = "word",
  upos = "upos", genre = NULL)
}
\value{
A dataframe
}
\description{
Function summarizes corpus, texts comprising corpus, and (optionally) genres ... by total tokens, unique tokens, and number of sentences.
}
