% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepText.R
\name{prepText}
\alias{prepText}
\alias{clr_prep_corpus}
\title{Prepare character string for annotation}
\usage{
clr_prep_corpus(x, hyphenate = TRUE)
}
\arguments{
\item{x}{A list of character strings}

\item{hyphenate}{if TRUE, trick spacy into preserving hyphenated words}
}
\value{
A list of character strings
}
\description{
Function cleans raw texts
}
