% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCorpus.R
\name{queryCorpus}
\alias{queryCorpus}
\alias{extractContext}
\alias{SimpleSearch}
\alias{GetContexts}
\alias{GetKeyPhrases}
\title{Get search in context}
\usage{
extractContext(x, search, LW, RW)

SimpleSearch(search, corp)

GetContexts(search, corp, LW, RW)

GetKeyPhrases(x, n = 5, aggBy = "lemma", flatten = TRUE, jitter = TRUE)
}
\arguments{
\item{search}{Gram/lexical pattern to be searched for}

\item{LW}{Size of context in number of words to left of the target}

\item{RW}{Size of context in number of words to right of the target}

\item{corp}{List of annotated texts to be searched}
}
\value{
A list of dataframes
}
\description{
These functions enable corpus search of gram constructions in context.
}
