% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_keyphrase.R
\name{search_keyphrase}
\alias{search_keyphrase}
\alias{clr_search_keyphrases}
\title{Extract key phrases from an annotated corpus.}
\usage{
clr_search_keyphrases(corp, n = 5, key_var = "lemma", flatten = TRUE,
  jitter = TRUE, remove_nums = TRUE, include = "doc_id")
}
\arguments{
\item{corp}{List of annotated texts to be searched}

\item{search}{Gram/lexical pattern to be searched for}

\item{LW}{Size of context in number of words to left of the target}

\item{RW}{Size of context in number of words to right of the target}
}
\value{
A list of dataframes
}
\description{
Function enable corpus search of gram constructions in context.
}
