% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_plus_context.R
\name{search_plus_context}
\alias{search_plus_context}
\alias{clr_extract_context}
\alias{clr_search_context}
\title{Get 'keyword-in-context' from an annotated corpus.}
\usage{
clr_extract_context(x, search, LW, RW)

clr_search_context(search, corp, LW, RW)
}
\arguments{
\item{search}{Gram/lexical pattern to be searched for}

\item{LW}{Size of context in number of words to left of the target}

\item{RW}{Size of context in number of words to right of the target}

\item{corp}{List of annotated texts to be searched}
}
\value{
A list of dataframes
}
\description{
Function(s) extract all instantiations of search pattern and NxN window of surrounding text from an annotated corpus.
}
