% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_simple.R
\name{search_simple}
\alias{search_simple}
\alias{clr_search_gramx}
\title{Get search in context}
\usage{
clr_search_gramx(search, corp)
}
\arguments{
\item{search}{Gram/lexical pattern to be searched for}

\item{corp}{List of annotated texts to be searched}
}
\value{
A list of dataframes
}
\description{
These functions enable corpus search of gram constructions in context.
}
