% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCorpus.R
\name{setCorpus}
\alias{setCorpus}
\alias{clr_prep_corpus}
\alias{clr_set_corpus}
\title{Modify text annotations for more robust corpus search}
\usage{
clr_prep_corpus(x, text_var = "text", hyphenate = TRUE)

clr_set_corpus(x, doc_var = "doc_id", token_var = "token",
  lemma_var = "lemma", tag_var = "tag", pos_var = "pos",
  sentence_var = "sentence_id", NER_as_tag = FALSE)
}
\arguments{
\item{x}{A list of dataframes}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
