% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_context.R
\name{summarize_context}
\alias{summarize_context}
\alias{clr_context_kwic}
\alias{clr_context_bow}
\title{Summarize results of corpuslingr::clr_search_contexts()}
\usage{
clr_context_kwic(x, include = c("doc_id", "lemma"))

clr_context_bow(x, content_only = TRUE, agg_var = c("lemma", "pos"))
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
These functions aggregate search results by frequency, BOW, and KWIC.
}
