% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSearch.R
\name{summarizeSearch}
\alias{summarizeSearch}
\alias{flatten_contexts}
\alias{corp_get_freq}
\alias{corp_context_kwic}
\alias{corp_context_bow}
\title{Summarize results of corpuslingr::GetContexts()}
\usage{
flatten_contexts(x)

corp_get_freq(x, agg_var = c("lemma", "token"))

corp_context_kwic(x, include = c("doc_id", "lemma"))

corp_context_bow(x, content_only = TRUE, agg_var = c("lemma", "pos"))
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
These functions aggregate search results by frequency, BOW, and KWIC.
}
