% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_cql.R
\docType{data}
\name{translate_cql}
\alias{translate_cql}
\alias{clr_build_search}
\alias{clr_ref_nounphrase}
\alias{clr_ref_keyphrase}
\alias{clr_cql_regex}
\title{Translate user search input to regex}
\format{An object of class \code{character} of length 1.}
\usage{
clr_build_search(x)

clr_ref_nounphrase

clr_ref_keyphrase

clr_cql_regex(x)
}
\arguments{
\item{x}{Gram/lexical pattern to be searched for}
}
\value{
A list of character strings
}
\description{
These functions convert CQL search to regex.
}
\keyword{datasets}
