% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateCQL.R
\docType{data}
\name{translateCQL}
\alias{translateCQL}
\alias{build_search}
\alias{corp_noun_phrase}
\alias{corp_key_phrase}
\alias{corp_cql_regex}
\title{Translate user search input to regex}
\format{An object of class \code{character} of length 1.}
\usage{
build_search(x)

corp_noun_phrase

corp_key_phrase

corp_cql_regex(x)
}
\arguments{
\item{x}{Gram/lexical pattern to be searched for}
}
\value{
A list of character strings
}
\description{
These functions convert CQL search to regex.
}
\keyword{datasets}
