% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpAnnotate.R
\name{corpAnnotate}
\alias{corpAnnotate}
\alias{buildTuple}
\alias{SpacyrAnnotate}
\title{Annotate a corpus of texts using the `spacyr` package}
\usage{
buildTuple(x)

SpacyrAnnotate(charList)
}
\arguments{
\item{charList}{A list of texts as character strings}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
