% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeCorpus.R
\name{describeCorpus}
\alias{describeCorpus}
\alias{clr_desc_corpus}
\title{Summarize results of corpuslingr::GetContexts()}
\usage{
clr_desc_corpus(x, doc = "id", sent = "sid", tok = "word",
  upos = "upos")
}
\value{
A dataframe
}
\description{
These functions aggregate search results by frequency, BOW, and KWIC.
}
