% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCorpus.R
\name{queryCorpus}
\alias{queryCorpus}
\alias{clr_extract_context}
\alias{clr_search_gramx}
\alias{clr_search_context}
\alias{clr_search_keyphrases}
\title{Get search in context}
\usage{
clr_extract_context(x, search, LW, RW)

clr_search_gramx(search, corp)

clr_search_context(search, corp, LW, RW)

clr_search_keyphrases(x, n = 5, key_var = "lemma", flatten = TRUE,
  jitter = TRUE)
}
\arguments{
\item{search}{Gram/lexical pattern to be searched for}

\item{LW}{Size of context in number of words to left of the target}

\item{RW}{Size of context in number of words to right of the target}

\item{corp}{List of annotated texts to be searched}
}
\value{
A list of dataframes
}
\description{
These functions enable corpus search of gram constructions in context.
}
