% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeWeb.R
\name{scrapeWeb}
\alias{scrapeWeb}
\alias{corp_web_gnews}
\alias{corp_web_scrape}
\title{Access GoogleNews quickly}
\usage{
corp_web_gnews(x, search = NULL, n = 30)

corp_web_scrape(y, link_var = "links")
}
\arguments{
\item{search}{Search topic for GoogleNews.  Defaults to NULL, which amounts to "Top Stories"}

\item{n}{Number of articles to get; max = 30.}
}
\value{
A dataframe of meta for articles collected.
}
\description{
These functions enable easy access to GoogleNews rss feed, and subsequent scraping.
}
