% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeWeb.R
\name{scrapeWeb}
\alias{scrapeWeb}
\alias{clr_web_gnews}
\alias{clr_web_scrape}
\title{Access GoogleNews quickly}
\usage{
clr_web_gnews(x, language = "en", country = "us", type = "topstories",
  search = NULL)

clr_web_scrape(y, link_var = "link")
}
\arguments{
\item{language}{Language locale code}

\item{country}{Country two letter code}

\item{type}{Search type: topic, topstories, term}

\item{search}{if type = term/topic, the topic or term to be searched.  Sets to NULL when type = topstories.}
}
\value{
A dataframe of meta for articles collected.
}
\description{
These functions enable easy access to GoogleNews rss feed, and subsequent scraping.
}
