% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCorpus.R
\name{setCorpus}
\alias{setCorpus}
\alias{SetNames}
\alias{SetTuple}
\alias{SetSearchCorpus}
\title{Modify text annotations for more robust corpus search}
\usage{
SetNames(x, old, new)

SetTuple(x)

SetSearchCorpus(x, doc_var = "doc_id", token_var = "token",
  lemma_var = "lemma", tag_var = "tag", pos_var = "pos",
  NER_as_tag = FALSE)
}
\arguments{
\item{x}{A list of dataframes}
}
\value{
A list of dataframes
}
\description{
These functions modify the output of `spacyr'
}
