% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSearch.R
\name{summarizeSearch}
\alias{summarizeSearch}
\alias{FlattenContexts}
\alias{GetSearchFreqs}
\alias{GetKWIC}
\alias{GetBOW}
\title{Summarize results of corpuslingr::GetContexts()}
\usage{
FlattenContexts(x)

GetSearchFreqs(x, aggBy = c("lemma", "targ"))

GetKWIC(x)

GetBOW(x, contentOnly = TRUE)
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
These functions aggregate search results by frequency, BOW, and KWIC.
}
