% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSearch.R
\name{summarizeSearch}
\alias{summarizeSearch}
\alias{clr_flatten_contexts}
\alias{clr_get_freq}
\alias{clr_context_kwic}
\alias{clr_context_bow}
\title{Summarize results of corpuslingr::GetContexts()}
\usage{
clr_flatten_contexts(x)

clr_get_freq(x, agg_var = c("lemma", "token"))

clr_context_kwic(x, include = c("doc_id", "lemma"))

clr_context_bow(x, content_only = TRUE, agg_var = c("lemma", "pos"))
}
\arguments{
\item{charList}{A dataframe}
}
\value{
A dataframes
}
\description{
These functions aggregate search results by frequency, BOW, and KWIC.
}
