% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateCQL.R
\docType{data}
\name{translateCQL}
\alias{translateCQL}
\alias{buildSearch}
\alias{nounPhrase}
\alias{keyPhrase}
\alias{CQLtoRegex}
\title{Translate user search input to regex}
\format{An object of class \code{character} of length 1.}
\usage{
buildSearch(x)

nounPhrase

keyPhrase

CQLtoRegex(x)
}
\arguments{
\item{x}{Gram/lexical pattern to be searched for}
}
\value{
A list of character strings
}
\description{
These functions convert CQL search to regex.
}
\keyword{datasets}
