% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateCQL.R
\docType{data}
\name{translateCQL}
\alias{translateCQL}
\alias{clr_build_search}
\alias{clr_nounphrase}
\alias{clr_keyphrase}
\alias{clr_cql_regex}
\title{Translate user search input to regex}
\format{An object of class \code{character} of length 1.}
\usage{
clr_build_search(x)

clr_nounphrase

clr_keyphrase

clr_cql_regex(x)
}
\arguments{
\item{x}{Gram/lexical pattern to be searched for}
}
\value{
A list of character strings
}
\description{
These functions convert CQL search to regex.
}
\keyword{datasets}
